Add chat UI, K8s scaling dashboard, docker-compose, DB schema, and complete documentation
Features:
- Self-contained chat page with 3 agents (Quiz, Weather, RAG), config panel, tool call display
- Live K8s scaling dashboard (polls /metrics + /k8s endpoints, shows HPA gauges, pods table)
- Gateway /metrics endpoint (service health checks, traffic stats) and /k8s endpoint (kubectl integration)
- In-memory MetricsCollector with thread-safe request tracking
- Production docker-compose.yml with all 11 containers (5 app services + 4 infra + frontend + ollama-init)
- Alembic migration 002: agents, chat_sessions, chat_messages, inference_costs tables with seed data
- K8s demo manifests (proxy pods + HPAs for gateway and agent-engine, 1-5 replicas)
- Load test script (10 concurrent users, configurable duration)

Documentation (7,700+ lines):
- tech-stack-complete-guide.md: All 16 components explained for fresh graduates
- kubernetes-scaling-guide.md: K8s concepts, HPA, minikube, scaling demo walkthrough
- claude-code-guide.md: How Claude Code built this app, CLAUDE.md strategy, debugging sessions
- architecture-analysis-2026.md: State-of-the-art analysis, FAANG comparison, 4.5/5 rating
- cloud-migration-guide.md: Deploy to GCP/Azure/AWS with docker-compose
- database-guide.md: PostgreSQL + pgvector, Redis VSS, MinIO, complete schema reference

Updated CLAUDE.md files (root, frontend, gateway, services, infra) to reflect current state.

+++ services/gateway/src/gateway/schema.py
@@ -60,7 +60,7 @@ class ChatMessage:
     content: str
     tool_calls: list[ToolCall] | None = None
     cost_usd: float | None = None
-    latency_ms: int | None = None
+    latency_ms: float | None = None
     created_at: datetime
